+++ app/controllers/banner/BannerTestArchiveController2.scala
@@ -1,24 +1,25 @@
-package controllers
+package controllers.banner
 
 import com.gu.googleauth.AuthAction
-import models.EpicTest
-import models.EpicTests._
+import controllers.S3ObjectsController
+import models.BannerTest
+import models.BannerTests._
 import play.api.libs.ws.WSClient
 import play.api.mvc.{AnyContent, ControllerComponents}
 import zio.DefaultRuntime
 
 import scala.concurrent.ExecutionContext
 
-class LiveblogEpicTestArchiveController(
+class BannerTestArchiveController2(
   authAction: AuthAction[AnyContent],
   components: ControllerComponents,
   ws: WSClient, stage: String,
   runtime: DefaultRuntime
-)(implicit ec: ExecutionContext) extends S3ObjectsController[EpicTest](
+)(implicit ec: ExecutionContext) extends S3ObjectsController[BannerTest](
   authAction,
   components,
   stage,
-  path = "archived-liveblog-epic-tests",
+  path = "archived-banner-tests2",
   nameGenerator = _.name,
   runtime
 )
