+++ app/controllers/banner/BannerTestsController.scala
@@ -1,6 +1,7 @@
-package controllers
+package controllers.banner
 
 import com.gu.googleauth.AuthAction
+import controllers.LockableS3ObjectController
 import models.BannerTests
 import play.api.libs.circe.Circe
 import play.api.libs.ws.WSClient
@@ -11,11 +12,11 @@ import zio.DefaultRuntime
 
 import scala.concurrent.ExecutionContext
 
-object BannerTestsController2 {
-  val name = "banner-tests2"
+object BannerTestsController {
+  val name = "banner-tests"
 }
 
-class BannerTestsController2(
+class BannerTestsController(
   authAction: AuthAction[AnyContent],
   components: ControllerComponents,
   ws: WSClient, stage: String,
@@ -24,14 +25,14 @@ class BannerTestsController2(
     authAction,
     components,
     stage,
-    name = BannerTestsController2.name,
+    name = BannerTestsController.name,
     dataObjectSettings = S3ObjectSettings(
       bucket = "gu-contributions-public",
-      key = s"banner/$stage/${BannerTestsController2.name}.json",
+      key = s"banner/$stage/${BannerTestsController.name}.json",
       publicRead = true,  // This data will be requested by dotcom
       cacheControl = Some("max-age=30"),
       surrogateControl = Some("max-age=86400")  // Cache for a day, and use cache purging after updates
     ),
-    fastlyPurger = FastlyPurger.fastlyPurger(stage, s"${BannerTestsController2.name}.json", ws),
+    fastlyPurger = FastlyPurger.fastlyPurger(stage, s"${BannerTestsController.name}.json", ws),
     runtime = runtime
   ) with Circe
