+++ app/controllers/epic/EpicTestsController.scala
@@ -1,13 +1,13 @@
-package controllers
+package controllers.epic
 
 import com.gu.googleauth.AuthAction
-import play.api.mvc._
+import controllers.LockableS3ObjectController
 import models.EpicTests
+import play.api.libs.circe.Circe
 import play.api.libs.ws.WSClient
+import play.api.mvc._
 import services.FastlyPurger
 import services.S3Client.S3ObjectSettings
-
-import play.api.libs.circe.Circe
 import zio.DefaultRuntime
 
 import scala.concurrent.ExecutionContext
